requirements.txt
pyinstaller==4.0
pillow
requests

